Change Doubles to BigDecimal and amount from int to long

+++ warehouse/core/src/main/java/core/CoreConst.java
@@ -4,8 +4,8 @@ package core;
  * Collection of constants that are used in the core module, and that might be changed in the future.
  */
 public abstract class CoreConst {
-  public static final int MIN_AMOUNT = 0;
-  public static final int MAX_AMOUNT = Integer.MAX_VALUE - 1;
+  public static final long MIN_AMOUNT = 0;
+  public static final long MAX_AMOUNT = Long.MAX_VALUE - 1;
 
   public static final double MIN_PRICE = 0;
   public static final double MAX_PRICE = Double.POSITIVE_INFINITY;
